extend to python scripting support
Current version is made for command  line only, i.e. command line arguments are taken from sys.argv.
I propose to add an argv argument to mainRoutine so that it can be called directly from a python session.
I have done it with sentinel2Stacked, but same could be done with others mainRoutines.